pyproject.toml
[project]
name = "hall-opt"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "elida", email = ""},
]
dependencies = [
    "matplotlib>=3.9.2",
    "juliacall>=0.9.23",
    "scipy>=1.14.1",
    "pandas>=2.2.3",
    "seaborn>=0.13.2",
    "arviz>=0.20.0",
    "sklearn>=0.0.post12",
    "scikit-learn>=1.5.2",
    "logging>=0.4.9.6",
    "datetime>=5.5",
    "nolds>=0.6.1",
    "MCMCIterators @ git+https://github.com/goroda/MCMCIterators.git",
    "pydantic>=2.10.5",
    "pyyaml>=6.0.2",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
